// src/models/Product.ts
import { Review } from "./Review";

type MetaData = {
  createdAt: string;
  updatedAt: string;
  barcode: string;
  qrCode: string;
};

type Dimensions = {
  width: number;
  height: number;
  depth: number;
};

export interface Product {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  tags: string[];
  brand: string;
  sku: string;
  weight: number;
  dimensions: Dimensions;
  warrantyInformation: string;
  shippingInformation: string;
  availabilityStatus: string;
  reviews: Review[];
  returnPolicy: string;
  minimumOrderQuantity: number;
  meta: MetaData;
  images: string[];
  thumbnail: string;
}